Renaming REST handlers to better align with the operations.
